Fix test_ref
I compeleted the FIXME part to make test_ref
can work normally.
The new way to implement test_ref is using a
2-dim array to store cities.

But original citits.txt is too large that will
make the program crush.So I add a new file
cities5000.txt , which is the reduced version
of the original file.